+++ DateTimeParts/DependencyInjection/DateTimeUtilsExtension.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace DigitalCraftsman\DateTimeUtils\DependencyInjection;
+namespace DigitalCraftsman\DateTimeParts\DependencyInjection;
 
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
