+++ DateTimeParts/Serializer/DateTimeNormalizer.php
@@ -2,27 +2,27 @@
 
 declare(strict_types=1);
 
-namespace DigitalCraftsman\DateTimeUtils\Serializer;
+namespace DigitalCraftsman\DateTimeParts\Serializer;
 
-use DigitalCraftsman\DateTimeUtils\ValueObject\Date;
+use DigitalCraftsman\DateTimeParts\ValueObject\DateTime;
 use Symfony\Component\Serializer\Normalizer\CacheableSupportsMethodInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
-final class DateNormalizer implements NormalizerInterface, DenormalizerInterface, CacheableSupportsMethodInterface
+final class DateTimeNormalizer implements NormalizerInterface, DenormalizerInterface, CacheableSupportsMethodInterface
 {
     public function supportsNormalization($data, $format = null, array $context = []): bool
     {
-        return $data instanceof Date;
+        return $data instanceof DateTime;
     }
 
     /** @param class-string $type */
     public function supportsDenormalization($data, $type, $format = null, array $context = []): bool
     {
-        return $type === Date::class;
+        return $type === DateTime::class;
     }
 
-    /** @param Date|null $object */
+    /** @param DateTime|null $object */
     public function normalize($object, $format = null, array $context = []): ?string
     {
         return $object === null
@@ -30,15 +30,14 @@ final class DateNormalizer implements NormalizerInterface, DenormalizerInterface
             : (string) $object;
     }
 
-    /** @param ?string $data */
-    public function denormalize($data, $type, $format = null, array $context = []): ?Date
+    /** @param string|null $data */
+    public function denormalize($data, $type, $format = null, array $context = []): ?DateTime
     {
         return $data === null
             ? null
-            : Date::fromString($data);
+            : DateTime::fromString($data);
     }
 
-    /** @codeCoverageIgnore */
     public function hasCacheableSupportsMethod(): bool
     {
         return true;
