+++ DateTimeParts/ValueObject/Year.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace DigitalCraftsman\DateTimeUtils\ValueObject;
+namespace DigitalCraftsman\DateTimeParts\ValueObject;
 
 /** @psalm-immutable */
 final class Year
